Splitting component to sub components
