+++ src/components/colorPicker/index.js
@@ -1,20 +1,15 @@
 import _ from 'lodash';
 import PropTypes from 'prop-types';
 import React, {PureComponent} from 'react';
-import {LayoutAnimation, StyleSheet, Keyboard, TextInput, PixelRatio, I18nManager} from 'react-native';
+import {StyleSheet} from 'react-native';
 import ColorPalette from './ColorPalette';
 import {SWATCH_MARGIN, SWATCH_SIZE} from './ColorSwatch';
-import {Constants} from '../../helpers';
 import {asBaseComponent} from '../../commons';
 import Assets from '../../assets';
-import {Colors, Typography} from '../../style';
+import {Colors} from '../../style';
 import View from '../view';
-import Text from '../text';
-import TouchableOpacity from '../touchableOpacity';
-import Dialog from '../dialog';
 import Button from '../button';
-import ColorSliderGroup from '../slider/ColorSliderGroup';
-import PanningProvider from '../panningViews/panningProvider';
+import ColorPickerDialog from './ColorPickerDialog';
 
 
 const ACCESSIBILITY_LABELS = {
@@ -23,7 +18,6 @@ const ACCESSIBILITY_LABELS = {
   doneButton: 'done',
   input: 'custom hex color code'
 };
-const KEYBOARD_HEIGHT = 216;
 
 /**
  * @description: A color picker component
@@ -85,80 +79,10 @@ class ColorPicker extends PureComponent {
   constructor(props) {
     super(props);
 
-    const color = Colors.getHSL(props.initialColor);
-    const text = this.getColorValue(props.initialColor);
-    const {valid} = this.getValidColorString(text);
-
     this.state = {
-      keyboardHeight: KEYBOARD_HEIGHT,
-      show: false,
-      color,
-      text,
-      valid
+      show: false
     };
   }
-
-  componentDidMount() {
-    this.keyboardDidShowListener = Keyboard.addListener('keyboardDidShow', this.keyboardDidShow);
-    this.keyboardDidHideListener = Keyboard.addListener('keyboardDidHide', this.keyboardDidHide);
-  }
-
-  componentWillUnmount() {
-    this.keyboardDidShowListener.remove();
-    this.keyboardDidHideListener.remove();
-  }
-
-  // TODO: remove method after migration to 'key' completed
-  UNSAFE_componentWillReceiveProps(nextProps) {
-    if (this.props.initialColor !== nextProps.initialColor) {
-      this.showMigrationMessage();
-
-      const text = this.getColorValue(nextProps.initialColor || this.props.initialColor);
-      const color = Colors.getHSL(nextProps.initialColor);
-      const {valid} = this.getValidColorString(text);
-
-      this.setState({color, text, valid});
-    }
-  }
-
-  showMigrationMessage() {
-    console.warn(`ColorPicker component will soon become uncontrolled component.` +
-    ` Thus 'initialColor' prop will become INITIAL, meaning, no updates to this prop will take effect` +
-    ` unless passing 'key' prop.`);
-  }
-
-  keyboardDidShow = e => {
-    if (Constants.isIOS && this.state.keyboardHeight !== e.endCoordinates.height) {
-      this.setState({keyboardHeight: e.endCoordinates.height});
-    }
-    // For down arrow button in Android keyboard
-    this.changeHeight(0);
-  };
-
-  keyboardDidHide = () => {
-    this.changeHeight(KEYBOARD_HEIGHT);
-  };
-
-  onFocus = () => {
-    this.changeHeight(0);
-  };
-
-  setFocus = () => {
-    if (this.textInput) {
-      this.textInput.focus();
-    }
-  };
-
-  changeHeight(height) {
-    if (Constants.isAndroid && this.state.keyboardHeight !== height) {
-      LayoutAnimation.configureNext(LayoutAnimation.Presets.easeInEaseOut);
-      this.setState({keyboardHeight: height});
-    }
-  }
-
-  showDialog = () => {
-    this.setState({show: true});
-  };
 
   get animatedIndex() {
     const {animatedIndex, colors} = this.props;
@@ -168,217 +92,17 @@ class ColorPicker extends PureComponent {
     return animatedIndex;
   }
 
-  getColorValue(color) {
-    if (!_.isString(color)) {
-      return;
-    }
-    return color.replace('#', '');
-  }
-
-  getHexColor(text) {
-    if (text && !Colors.isTransparent(text)) {
-      const trimmed = text.replace(/\s+/g, '');
-      const hex = `#${trimmed}`;
-      return hex;
-    }
-    return text;
-  }
-
-  getHexString(color) {
-    return _.toUpper(Colors.getHexString(color));
-  }
-
-  getTextColor(color) {
-    return Colors.isDark(color) ? Colors.white : Colors.dark10;
-  }
-
-  getValidColorString(text) {
-    const hex = this.getHexColor(text);
-
-    if (Colors.isValidHex(hex)) {
-      return {hex, valid: true};
-    }
-    return {undefined, valid: false};
-  }
-
-  applyColor = (text) => {
-    const {hex, valid} = this.getValidColorString(text);
-
-    if (hex) {
-      this.setState({color: Colors.getHSL(hex), text, valid});
-    } else {
-      this.setState({text, valid});
-    }
+  showDialog = () => {
+    this.setState({show: true});
   };
 
-  updateColor(color) {
-    const hex = this.getHexString(color);
-    const text = this.getColorValue(hex);
-    this.setState({color, text, valid: true});
-  }
-
-  resetValues() {
-    const color = Colors.getHSL(this.props.initialColor);
-    const text = this.getColorValue(this.props.initialColor);
-    const {valid} = this.getValidColorString(text);
-
-    this.setState({
-      show: false,
-      color,
-      text,
-      valid
-    });
-  }
-
-  onSliderValueChange = color => {
-    const c = Colors.getHSL(color);
-    this.updateColor(c);
-  };
-
-  onChangeText = value => {
-    this.applyColor(value);
-  };
-
-  onDonePressed = () => {
-    const {text} = this.state;
-    const {hex} = this.getValidColorString(text);
-
-    if (hex) {
-      _.invoke(this.props, 'onSubmit', hex, this.getTextColor(hex));
-      this.onDismiss();
-    }
-  };
-
-  onDismiss = () => {
-    this.resetValues();
-    _.invoke(this.props, 'onDismiss');
-  };
-
-  renderHeader() {
-    const {useCustomTheme, accessibilityLabels} = this.props;
-    const {valid} = this.state;
-
-    return (
-      <View row spread bg-white paddingH-20 style={styles.header}>
-        <Button
-          link
-          iconSource={Assets.icons.x}
-          iconStyle={{tintColor: Colors.dark10}}
-          onPress={this.onDismiss}
-          accessibilityLabel={accessibilityLabels.dismissButton}
-        />
-        <Button
-          useCustomTheme={useCustomTheme}
-          disabled={!valid}
-          link
-          iconSource={Assets.icons.check}
-          onPress={this.onDonePressed}
-          accessibilityLabel={accessibilityLabels.doneButton}
-        />
-      </View>
-    );
-  }
-
-  renderSliders() {
-    const {keyboardHeight, color} = this.state;
-    const colorValue = color.a === 0 ? Colors.black : Colors.getHexString(color);
-
-    return (
-      <ColorSliderGroup
-        initialColor={colorValue}
-        containerStyle={[styles.sliderGroup, {height: keyboardHeight}]}
-        sliderContainerStyle={styles.slider}
-        showLabels
-        labelsStyle={styles.label}
-        onValueChange={this.onSliderValueChange}
-        accessible={false}
-      />
-    );
-  }
-
-  renderPreview() {
-    const {accessibilityLabels, previewInputStyle} = this.props;
-    const {color, text} = this.state;
-    const hex = this.getHexString(color);
-    const textColor = this.getTextColor(hex);
-    const fontScale = PixelRatio.getFontScale();
-    const value = Colors.isTransparent(text) ? '000000' : text;
-
-    return (
-      <View style={[styles.preview, {backgroundColor: hex}]}>
-        <TouchableOpacity center onPress={this.setFocus} activeOpacity={1} accessible={false}>
-          <View style={styles.inputContainer}>
-            <Text
-              text60
-              white
-              marginL-13
-              marginR-5={Constants.isIOS}
-              style={{
-                color: textColor,
-                transform: [{scaleX: I18nManager.isRTL ? -1 : 1}]
-              }}
-              accessible={false}
-            >
-              #
-            </Text>
-            <TextInput
-              ref={r => (this.textInput = r)}
-              value={value}
-              maxLength={6}
-              numberOfLines={1}
-              onChangeText={this.onChangeText}
-              style={[
-                styles.input,
-                {color: textColor, width: (value.length + 1) * 16.5 * fontScale},
-                Constants.isAndroid && {padding: 0},
-                previewInputStyle
-              ]}
-              selectionColor={textColor}
-              underlineColorAndroid="transparent"
-              autoCorrect={false}
-              autoComplete={'off'}
-              autoCapitalize={'characters'}
-              // keyboardType={'numbers-and-punctuation'} // doesn't work with `autoCapitalize`
-              returnKeyType={'done'}
-              enablesReturnKeyAutomatically
-              onFocus={this.onFocus}
-              accessibilityLabel={accessibilityLabels.input}
-            />
-          </View>
-          <View style={[{backgroundColor: textColor}, styles.underline]}/>
-        </TouchableOpacity>
-      </View>
-    );
-  }
-
-  renderDialog() {
-    const {testID, dialogProps} = this.props;
-    const {show} = this.state;
-
-    return (
-      <Dialog
-        migrate
-        visible={show}
-        width="100%"
-        height={null}
-        bottom
-        centerH
-        onDismiss={this.onDismiss}
-        containerStyle={styles.dialog}
-        panDirection={PanningProvider.Directions.DOWN}
-        testID={`${testID}.dialog`}
-        supportedOrientations={['portrait', 'landscape', 'landscape-left', 'landscape-right']} // iOS only
-        {...dialogProps}
-      >
-        {this.renderHeader()}
-        {this.renderPreview()}
-        {this.renderSliders()}
-      </Dialog>
-    );
+  hideDialog = () => {
+    this.setState({show: false});
   }
 
   render() {
-    const {colors, value, testID, accessibilityLabels} = this.props;
+    const {/* initialColor,  */colors, value, testID, accessibilityLabels} = this.props;
+    const {show} = this.state;
 
     return (
       <View row testID={testID}>
@@ -404,7 +128,12 @@ class ColorPicker extends PureComponent {
             accessibilityLabel={accessibilityLabels.addButton}
           />
         </View>
-        {this.renderDialog()}
+        <ColorPickerDialog
+          {...this.props}
+          visible={show}
+          onDismiss={this.hideDialog}
+          /* key={initialColor} */
+        />
       </View>
     );
   }
@@ -418,18 +147,12 @@ class ColorPicker extends PureComponent {
 export default asBaseComponent(ColorPicker);
 
 
-const BORDER_RADIUS = 12;
 const plusButtonContainerWidth = SWATCH_SIZE + 20 + 12;
 const plusButtonContainerHeight = 92 - 2 * SWATCH_MARGIN;
 
 const styles = StyleSheet.create({
   palette: {
     paddingLeft: plusButtonContainerWidth
-  },
-  dialog: {
-    backgroundColor: Colors.white,
-    borderTopLeftRadius: BORDER_RADIUS,
-    borderTopRightRadius: BORDER_RADIUS
   },
   buttonContainer: {
     position: 'absolute',
@@ -447,43 +170,5 @@ const styles = StyleSheet.create({
     width: SWATCH_SIZE,
     height: SWATCH_SIZE,
     marginRight: 12
-  },
-  preview: {
-    height: 200,
-    alignItems: 'center',
-    justifyContent: 'center'
-  },
-  header: {
-    height: 56,
-    borderTopLeftRadius: BORDER_RADIUS,
-    borderTopRightRadius: BORDER_RADIUS
-  },
-  inputContainer: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    flexDirection: 'row',
-    marginBottom: Constants.isAndroid ? 5 : 8,
-    transform: [{scaleX: I18nManager.isRTL ? -1 : 1}]
-  },
-  input: {
-    ...Typography.text60,
-    letterSpacing: 3,
-    transform: [{scaleX: I18nManager.isRTL ? -1 : 1}]
-  },
-  underline: {
-    height: 1.5,
-    width: Constants.isAndroid ? 119 : 134,
-    marginRight: Constants.isAndroid ? 13 : 8
-  },
-  sliderGroup: {
-    paddingTop: 12,
-    marginHorizontal: 20
-  },
-  slider: {
-    marginBottom: 15,
-    height: 26
-  },
-  label: {
-    marginBottom: 3
   }
 });
